requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.12.0
numpy>=1.24.0
Pillow>=9.0.0
pytest>=7.0.0
